[REFACTOR] Move content to unified priv/content directory structure
- Move priv/posts/ -> priv/content/archived/
- Move priv/assets/ -> priv/content/assets/
- Move priv/projects/ -> priv/content/projects/
- Update source paths in Post, Asset, Project schemas
- Update mix blog.gen.post task path
- Update all test fixture paths to match new structure

This establishes the unified content directory structure for the
zettelkasten redesign. All existing content now lives under priv/content/
with type-specific subdirectories.
