+++ apps/blog/lib/blog/posts/post.ex
@@ -3,7 +3,7 @@ defmodule Blog.Posts.Post do
   use Blog.Schema
 
   use Blog.Resource,
-    source_dir: "priv/posts",
+    source_dir: "priv/content/archived",
     preprocess: &Blog.Tags.label_to_id/1,
     import: &Blog.Posts.upsert_post/1
