+++ apps/blog/lib/blog/projects/project.ex
@@ -3,7 +3,7 @@ defmodule Blog.Projects.Project do
   use Blog.Schema
 
   use Blog.Resource,
-    source_dir: "priv/projects",
+    source_dir: "priv/content/projects",
     preprocess: &Blog.Tags.label_to_id/1,
     import: &Blog.Projects.upsert_project/1
