+++ apps/blog/priv/content/archived/20230722200932_nixos_kool_aid.md
