+++ apps/blog/priv/content/archived/20230916100331_elixir_pipes.md
