+++ apps/blog/priv/content/archived/20240904100331_story_of_a_function.md
