+++ apps/blog/priv/content/archived/20241126141959_dear_friend_you_built_an_erlang.md
