+++ apps/blog/priv/content/archived/20250521144307_my_take_on_consciousness.md
