+++ apps/blog/priv/content/archived/20250624122039_ny_take_on_free_will.md
