+++ apps/blog/priv/content/archived/20250630140400_simple_elixir_optimization.md
