+++ apps/blog/priv/content/archived/20250812142141_a_nixoswindows_11_dev_env.md
