+++ apps/blog/priv/content/assets/avatar.jpg
