+++ apps/blog/priv/content/assets/banner.png
