+++ apps/blog/priv/content/assets/blog_2.png
