+++ apps/blog/priv/content/assets/rufus.jpg
