+++ apps/blog/priv/content/projects/projects.yaml
