+++ apps/blog/test/blog/assets/asset_test.exs
@@ -6,22 +6,22 @@ defmodule Blog.Assets.AssetTest do
 
   @test_image_path Path.join([
                      File.cwd!(),
-                     "test/fixtures/priv/assets/test_image.jpg"
+                     "test/fixtures/priv/content/assets/test_image.jpg"
                    ])
 
   @test_text_path Path.join([
                     File.cwd!(),
-                    "test/fixtures/priv/assets/test_document.txt"
+                    "test/fixtures/priv/content/assets/test_document.txt"
                   ])
 
   @test_pdf_path Path.join([
                    File.cwd!(),
-                   "test/fixtures/priv/assets/test_document.pdf"
+                   "test/fixtures/priv/content/assets/test_document.pdf"
                  ])
 
   @invalid_image_path Path.join([
                         File.cwd!(),
-                        "test/fixtures/priv/assets/invalid_image.jpg"
+                        "test/fixtures/priv/content/assets/invalid_image.jpg"
                       ])
 
   describe "changeset/2 - validation" do
