+++ apps/blog/test/blog/assets_test.exs
@@ -6,7 +6,7 @@ defmodule Blog.AssetsTest do
 
   @test_image_path Path.join([
                      File.cwd!(),
-                     "test/fixtures/priv/assets/test_image.jpg"
+                     "test/fixtures/priv/content/assets/test_image.jpg"
                    ])
 
   describe "list_assets/1" do
