+++ apps/blog/test/fixtures/priv/content/assets/invalid_image.jpg
