+++ apps/blog/test/fixtures/priv/content/assets/test_image.jpg
