+++ apps/blog_web/test/blog_web/controllers/assets_controller_test.exs
@@ -5,7 +5,7 @@ defmodule BlogWeb.AssetsControllerTest do
 
   @test_image_path Path.join([
                      File.cwd!(),
-                     "test/fixtures/priv/assets/test_image.jpg"
+                     "test/fixtures/priv/content/assets/test_image.jpg"
                    ])
 
   describe "GET /assets/images/:name" do
